+++ zerver/lib/outgoing_webhook.py
@@ -10,7 +10,7 @@ from django.utils.translation import gettext as _
 from requests import Response
 
 from version import ZULIP_VERSION
-from zerver.lib.actions import check_send_message
+from zerver.actions.message_send import check_send_message
 from zerver.lib.exceptions import JsonableError
 from zerver.lib.message import MessageDict
 from zerver.lib.outgoing_http import OutgoingSession
