+++ zerver/tests/test_mirror_users.py
@@ -4,7 +4,7 @@ from unittest import mock
 from django.db import IntegrityError
 from django.utils.timezone import now as timezone_now
 
-from zerver.lib.actions import create_mirror_user_if_needed
+from zerver.actions.message_send import create_mirror_user_if_needed
 from zerver.lib.create_user import create_user_profile
 from zerver.lib.test_classes import ZulipTestCase
 from zerver.lib.test_helpers import reset_emails_in_zulip_realm
@@ -172,7 +172,7 @@ class MirroredMessageUsersTest(ZulipTestCase):
             create_user_profile(**kwargs).save()
             raise IntegrityError()
 
-        with mock.patch("zerver.lib.actions.create_user", side_effect=create_user) as m:
+        with mock.patch("zerver.actions.message_send.create_user", side_effect=create_user) as m:
             mirror_fred_user = create_mirror_user_if_needed(
                 realm,
                 email,
