+++ zerver/tests/test_tutorial.py
@@ -1,6 +1,6 @@
 from django.conf import settings
 
-from zerver.lib.actions import internal_send_private_message
+from zerver.actions.message_send import internal_send_private_message
 from zerver.lib.test_classes import ZulipTestCase
 from zerver.lib.test_helpers import message_stream_count, most_recent_message
 from zerver.models import UserProfile, get_system_bot
