+++ zerver/webhooks/dialogflow/view.py
@@ -3,8 +3,8 @@ from typing import Any, Dict
 
 from django.http import HttpRequest, HttpResponse
 
+from zerver.actions.message_send import check_send_private_message
 from zerver.decorator import webhook_view
-from zerver.lib.actions import check_send_private_message
 from zerver.lib.request import REQ, RequestNotes, has_request_variables
 from zerver.lib.response import json_success
 from zerver.models import UserProfile, get_user
