+++ zerver/webhooks/slack/view.py
@@ -2,8 +2,8 @@ from django.http import HttpRequest
 from django.http.response import HttpResponse
 from django.utils.translation import gettext as _
 
+from zerver.actions.message_send import check_send_stream_message
 from zerver.decorator import webhook_view
-from zerver.lib.actions import check_send_stream_message
 from zerver.lib.exceptions import JsonableError
 from zerver.lib.request import REQ, RequestNotes, has_request_variables
 from zerver.lib.response import json_success
