+++ zerver/webhooks/teamcity/view.py
@@ -5,11 +5,11 @@ from typing import Any, Dict, List, Optional
 from django.db.models import Q
 from django.http import HttpRequest, HttpResponse
 
-from zerver.decorator import webhook_view
-from zerver.lib.actions import (
+from zerver.actions.message_send import (
     check_send_private_message,
     send_rate_limited_pm_notification_to_bot_owner,
 )
+from zerver.decorator import webhook_view
 from zerver.lib.request import REQ, RequestNotes, has_request_variables
 from zerver.lib.response import json_success
 from zerver.lib.send_email import FromAddress
