+++ zerver/webhooks/uptimerobot/view.py
@@ -3,8 +3,8 @@ from typing import Any, Dict
 from django.http import HttpRequest, HttpResponse
 from django.utils.translation import gettext as _
 
+from zerver.actions.message_send import send_rate_limited_pm_notification_to_bot_owner
 from zerver.decorator import webhook_view
-from zerver.lib.actions import send_rate_limited_pm_notification_to_bot_owner
 from zerver.lib.exceptions import JsonableError
 from zerver.lib.request import REQ, has_request_variables
 from zerver.lib.response import json_success
